+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+- Add projectname in parentheses to search result on todos/tickets
+- Expand filter results by tolowercase'ing, because the result from the leantime has type set to be either "Task" and "task". 
+- remove sprintname/client from searcher to simplify result
+
 ## [1.2.1] 2024-04-29
 
 - Used `APP_ROOT` for installing assets.